pyproject.toml
[project]
name = "2cent-tts-onnx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "espeakng-loader>=0.2.4",
    "gguf>=0.16.3",
    "phonemizer-fork>=3.3.2",
    "snac>=1.2.1",
    "tokenizers>=0.21.1",
    "torch>=2.7.0",
    "transformers>=4.52.3",
]
